Extract reusable parts of output widget
